Refs #3550. Taken common factor for client management in SerialServer.
