+++ CMakeLists.txt
@@ -115,6 +115,7 @@ set(SRCS
     src/cpp/transport/Server.cpp
     src/cpp/transport/udp/UDPServerBase.cpp
     src/cpp/transport/tcp/TCPServerBase.cpp
+    src/cpp/transport/serial/SerialServerBase.cpp
     src/cpp/transport/serial/serial_protocol.c
     ${TRANSPORT_SRCS}
     )
