+++ include/uxr/agent/transport/serial/SerialServerLinux.hpp
@@ -15,10 +15,8 @@
 #ifndef _UXR_AGENT_TRANSPORT_SERIAL_SERVER_HPP_
 #define _UXR_AGENT_TRANSPORT_SERIAL_SERVER_HPP_
 
-#include <uxr/agent/transport/Server.hpp>
-#include <uxr/agent/transport/serial/serial_protocol.h>
-#include <uxr/agent/config.hpp>
-#include <unordered_map>
+#include <uxr/agent/transport/serial/SerialServerBase.hpp>
+
 #include <cstdint>
 #include <cstddef>
 #include <sys/poll.h>
@@ -26,53 +24,26 @@
 namespace eprosima {
 namespace uxr {
 
-/**************************************************************************************************
- * Serial EndPoint.
- **************************************************************************************************/
-class SerialEndPoint : public EndPoint
-{
-public:
-    SerialEndPoint(uint8_t addr) { addr_ = addr; }
-    ~SerialEndPoint() {}
-
-    uint8_t get_addr() const { return addr_; }
-
-public:
-    uint8_t addr_;
-};
-
-/**************************************************************************************************
- * Serial Server.
- **************************************************************************************************/
-class SerialServer : public Server
+class SerialServer : public SerialServerBase
 {
 public:
     SerialServer(int fd, uint8_t addr);
     ~SerialServer() = default;
 
-    virtual void on_create_client(EndPoint* source, const dds::xrce::ClientKey& client_key) override;
-    virtual void on_delete_client(EndPoint* source) override;
-    virtual const dds::xrce::ClientKey get_client_key(EndPoint* source) override;
-    virtual std::unique_ptr<EndPoint> get_source(const dds::xrce::ClientKey& client_key) override;
-
 private:
-    virtual bool init() override;
-    virtual bool close() override;
-    virtual bool recv_message(InputPacket& input_packet, int timeout) override;
-    virtual bool send_message(OutputPacket output_packet) override;
-    virtual int get_error() override;
+    bool init() override;
+    bool close() override;
+    bool recv_message(InputPacket& input_packet, int timeout) override;
+    bool send_message(OutputPacket output_packet) override;
+    int get_error() override;
     static size_t write_data(void* instance, uint8_t* buf, size_t len);
     static size_t read_data(void* instance, uint8_t* buf, size_t len, int timeout);
 
 private:
-    uint8_t addr_;
     struct pollfd poll_fd_;
     uint8_t buffer_[SERIAL_TRANSPORT_MTU];
     uxrSerialIO serial_io_;
     int errno_;
-    std::unordered_map<uint8_t, uint32_t> source_to_client_map_;
-    std::unordered_map<uint32_t, uint8_t> client_to_source_map_;
-    std::mutex clients_mtx_;
 };
 
 } // namespace uxr
