+++ src/cpp/transport/serial/SerialServerLinux.cpp
@@ -19,94 +19,13 @@ namespace eprosima {
 namespace uxr {
 
 SerialServer::SerialServer(int fd, uint8_t addr)
-    : addr_(addr),
+    : SerialServerBase(addr),
       poll_fd_(),
       buffer_{0},
       serial_io_(),
-      errno_(0),
-      source_to_client_map_{},
-      client_to_source_map_{}
+      errno_(0)
 {
     poll_fd_.fd = fd;
-}
-
-void SerialServer::on_create_client(EndPoint* source, const dds::xrce::ClientKey& client_key)
-{
-    SerialEndPoint* endpoint = static_cast<SerialEndPoint*>(source);
-    uint8_t source_id = endpoint->get_addr();
-    uint32_t client_id = client_key.at(0) + (client_key.at(1) << 8) + (client_key.at(2) << 16) + (client_key.at(3) <<24);
-
-    /* Update maps. */
-    std::lock_guard<std::mutex> lock(clients_mtx_);
-    auto it_client = client_to_source_map_.find(client_id);
-    if (it_client != client_to_source_map_.end())
-    {
-        source_to_client_map_.erase(it_client->second);
-        it_client->second = source_id;
-    }
-    else
-    {
-        client_to_source_map_.insert(std::make_pair(client_id, source_id));
-    }
-
-    auto it_source = source_to_client_map_.find(source_id);
-    if (it_source != source_to_client_map_.end())
-    {
-        it_source->second = client_id;
-    }
-    else
-    {
-        source_to_client_map_.insert(std::make_pair(source_id, client_id));
-    }
-}
-
-void SerialServer::on_delete_client(EndPoint* source)
-{
-    SerialEndPoint* endpoint = static_cast<SerialEndPoint*>(source);
-    uint8_t source_id = endpoint->get_addr();
-
-    /* Update maps. */
-    std::lock_guard<std::mutex> lock(clients_mtx_);
-    auto it = source_to_client_map_.find(source_id);
-    if (it != source_to_client_map_.end())
-    {
-        client_to_source_map_.erase(it->second);
-        source_to_client_map_.erase(it->first);
-    }
-}
-
-const dds::xrce::ClientKey SerialServer::get_client_key(EndPoint* source)
-{
-    dds::xrce::ClientKey client_key;
-    SerialEndPoint* endpoint = static_cast<SerialEndPoint*>(source);
-    std::lock_guard<std::mutex> lock(clients_mtx_);
-    auto it = source_to_client_map_.find(endpoint->get_addr());
-    if (it != source_to_client_map_.end())
-    {
-        client_key.at(0) = it->second & 0x000000FF;
-        client_key.at(1) = (it->second & 0x0000FF00) >> 8;
-        client_key.at(2) = (it->second & 0x00FF0000) >> 16;
-        client_key.at(3) = (it->second & 0xFF000000) >> 24;
-    }
-    else
-    {
-        client_key = dds::xrce::CLIENTKEY_INVALID;
-    }
-    return client_key;
-}
-
-std::unique_ptr<EndPoint> SerialServer::get_source(const dds::xrce::ClientKey& client_key)
-{
-    std::unique_ptr<EndPoint> source;
-    uint32_t client_id = client_key.at(0) + (client_key.at(1) << 8) + (client_key.at(2) << 16) + (client_key.at(3) <<24);
-    std::lock_guard<std::mutex> lock(clients_mtx_);
-    auto it = client_to_source_map_.find(client_id);
-    if (it != client_to_source_map_.end())
-    {
-        uint64_t source_id = it->second;
-        source.reset(new SerialEndPoint(source_id));
-    }
-    return source;
 }
 
 bool SerialServer::init()
@@ -123,53 +42,6 @@ bool SerialServer::init()
 bool SerialServer::close()
 {
     return 0 == ::close(poll_fd_.fd);
-}
-
-bool SerialServer::recv_message(InputPacket& input_packet, int timeout)
-{
-    bool rv = true;
-    uint8_t remote_addr;
-    size_t bytes_read = uxr_read_serial_msg(&serial_io_,
-                                            read_data,
-                                            this,
-                                            buffer_,
-                                            sizeof(buffer_),
-                                            &remote_addr,
-                                            timeout);
-    if (0 < bytes_read)
-    {
-        input_packet.message.reset(new InputMessage(buffer_, static_cast<size_t>(bytes_read)));
-        input_packet.source.reset(new SerialEndPoint(remote_addr));
-    }
-    else
-    {
-        errno_ = -1;
-        rv = false;
-    }
-    return rv;
-}
-
-bool SerialServer::send_message(OutputPacket output_packet)
-{
-    bool rv = false;
-    const SerialEndPoint* destination = static_cast<const SerialEndPoint*>(output_packet.destination.get());
-    size_t bytes_written = uxr_write_serial_msg(&serial_io_,
-                                                write_data,
-                                                this,
-                                                output_packet.message->get_buf(),
-                                                output_packet.message->get_len(),
-                                                destination->get_addr());
-    if ((0 < bytes_written) && (bytes_written == output_packet.message->get_len()))
-    {
-        rv = true;
-    }
-    errno_ = rv ? 0 : -1;
-    return rv;
-}
-
-int SerialServer::get_error()
-{
-    return errno_;
 }
 
 size_t SerialServer::write_data(void* instance, uint8_t* buf, size_t len)
@@ -198,6 +70,53 @@ size_t SerialServer::read_data(void* instance, uint8_t* buf, size_t len, int tim
         }
     }
     return rv;
+}
+
+bool SerialServer::recv_message(InputPacket& input_packet, int timeout)
+{
+    bool rv = false;
+    uint8_t remote_addr;
+    size_t bytes_read = uxr_read_serial_msg(&serial_io_,
+                                            read_data,
+                                            this,
+                                            buffer_,
+                                            sizeof(buffer_),
+                                            &remote_addr,
+                                            timeout);
+    if (0 < bytes_read)
+    {
+        input_packet.message.reset(new InputMessage(buffer_, static_cast<size_t>(bytes_read)));
+        input_packet.source.reset(new SerialEndPoint(remote_addr));
+        rv = true;
+    }
+    else
+    {
+        errno_ = -1;
+    }
+    return rv;
+}
+
+bool SerialServer::send_message(OutputPacket output_packet)
+{
+    bool rv = false;
+    const SerialEndPoint* destination = static_cast<const SerialEndPoint*>(output_packet.destination.get());
+    size_t bytes_written = uxr_write_serial_msg(&serial_io_,
+                                                write_data,
+                                                this,
+                                                output_packet.message->get_buf(),
+                                                output_packet.message->get_len(),
+                                                destination->get_addr());
+    if ((0 < bytes_written) && (bytes_written == output_packet.message->get_len()))
+    {
+        rv = true;
+    }
+    errno_ = rv ? 0 : -1;
+    return rv;
+}
+
+int SerialServer::get_error()
+{
+    return errno_;
 }
 
 } // namespace uxr
